feat: greengrass.service.template: add After=systemd-time-wait-sync.service
This will allow if systemd-time-wait-sync.service is enabled to wait for a
synced clock. If this is not enabled it will not wait.
Especially in systems without a RTC (like RPi) it will take some time after boot
till NTP will sync clock and if Greengrass is already started this time jump will
cause issues.